+++ fluent-react/examples/redux-sync/src/l10n.js
@@ -12,22 +12,22 @@ class AppLocalizationProvider extends Component {
   }
 
   render() {
-    const { messages, children } = this.props;
+    const { bundles, children } = this.props;
 
-    if (!messages) {
+    if (!bundles) {
       // Show a loader
       return <div>…</div>;
     }
 
     return (
-      <LocalizationProvider messages={messages}>
+      <LocalizationProvider bundles={bundles}>
         {children}
       </LocalizationProvider>
     );
   }
 }
 
-const mapStateToProps = state => ({ messages: state.messages });
+const mapStateToProps = state => ({ bundles: state.bundles });
 const mapDispatchToProps = { changeLocales };
 
 export default connect(mapStateToProps, mapDispatchToProps)(