+++ fluent-sequence/src/index.js
@@ -1,7 +1,7 @@
 /*
  * @module fluent-sequence
- * @overview Manage ordered sequences of MessageContexts.
+ * @overview Manage ordered sequences of FluentBundles.
  */
 
-export {default as mapContextSync} from "./map_sync";
-export {default as mapContextAsync} from "./map_async";
+export {default as mapBundleSync} from "./map_sync";
+export {default as mapBundleAsync} from "./map_async";