+++ fluent-sequence/test/fallback_sync_test.js
@@ -1,87 +1,87 @@
 import assert from 'assert';
 
 import {CachedSyncIterable} from 'cached-iterable';
-import MessageContext from './message_context_stub';
-import {mapContextSync} from '../src/index';
+import FluentBundle from './bundle_stub';
+import {mapBundleSync} from '../src/index';
 
 suite('Sync Fallback — single id', function() {
-  let ctx1, ctx2;
+  let bundle1, bundle2;
 
   suiteSetup(function() {
-    ctx1 = new MessageContext();
-    ctx1._setMessages(['bar']);
-    ctx2 = new MessageContext();
-    ctx2._setMessages(['foo', 'bar']);
+    bundle1 = new FluentBundle();
+    bundle1._setMessages(['bar']);
+    bundle2 = new FluentBundle();
+    bundle2._setMessages(['foo', 'bar']);
   });
 
   test('eager iterable', function() {
-    const contexts = new CachedSyncIterable([ctx1, ctx2]);
-    assert.equal(mapContextSync(contexts, 'foo'), ctx2);
-    assert.equal(mapContextSync(contexts, 'bar'), ctx1);
+    const bundles = new CachedSyncIterable([bundle1, bundle2]);
+    assert.equal(mapBundleSync(bundles, 'foo'), bundle2);
+    assert.equal(mapBundleSync(bundles, 'bar'), bundle1);
   });
 
   test('eager iterable works more than once', function() {
-    const contexts = new CachedSyncIterable([ctx1, ctx2]);
-    assert.equal(mapContextSync(contexts, 'foo'), ctx2);
-    assert.equal(mapContextSync(contexts, 'bar'), ctx1);
-    assert.equal(mapContextSync(contexts, 'foo'), ctx2);
-    assert.equal(mapContextSync(contexts, 'bar'), ctx1);
+    const bundles = new CachedSyncIterable([bundle1, bundle2]);
+    assert.equal(mapBundleSync(bundles, 'foo'), bundle2);
+    assert.equal(mapBundleSync(bundles, 'bar'), bundle1);
+    assert.equal(mapBundleSync(bundles, 'foo'), bundle2);
+    assert.equal(mapBundleSync(bundles, 'bar'), bundle1);
   });
 
   test('lazy iterable', function() {
-    function *generateMessages() {
-      yield *[ctx1, ctx2];
+    function *generateBundles() {
+      yield *[bundle1, bundle2];
     }
 
-    const contexts = new CachedSyncIterable(generateMessages());
-    assert.equal(mapContextSync(contexts, 'foo'), ctx2);
-    assert.equal(mapContextSync(contexts, 'bar'), ctx1);
+    const bundles = new CachedSyncIterable(generateBundles());
+    assert.equal(mapBundleSync(bundles, 'foo'), bundle2);
+    assert.equal(mapBundleSync(bundles, 'bar'), bundle1);
   });
 
   test('lazy iterable works more than once', function() {
-    function *generateMessages() {
-      yield *[ctx1, ctx2];
+    function *generateBundles() {
+      yield *[bundle1, bundle2];
     }
 
-    const contexts = new CachedSyncIterable(generateMessages());
-    assert.equal(mapContextSync(contexts, 'foo'), ctx2);
-    assert.equal(mapContextSync(contexts, 'bar'), ctx1);
-    assert.equal(mapContextSync(contexts, 'foo'), ctx2);
-    assert.equal(mapContextSync(contexts, 'bar'), ctx1);
+    const bundles = new CachedSyncIterable(generateBundles());
+    assert.equal(mapBundleSync(bundles, 'foo'), bundle2);
+    assert.equal(mapBundleSync(bundles, 'bar'), bundle1);
+    assert.equal(mapBundleSync(bundles, 'foo'), bundle2);
+    assert.equal(mapBundleSync(bundles, 'bar'), bundle1);
   });
 });
 
 suite('Sync Fallback — multiple ids', function() {
-  let ctx1, ctx2;
+  let bundle1, bundle2;
 
   suiteSetup(function() {
-    ctx1 = new MessageContext();
-    ctx1._setMessages(['foo', 'bar']);
-    ctx2 = new MessageContext();
-    ctx2._setMessages(['foo', 'bar', 'baz']);
+    bundle1 = new FluentBundle();
+    bundle1._setMessages(['foo', 'bar']);
+    bundle2 = new FluentBundle();
+    bundle2._setMessages(['foo', 'bar', 'baz']);
   });
 
   test('existing translations', function() {
-    const contexts = new CachedSyncIterable([ctx1, ctx2]);
+    const bundles = new CachedSyncIterable([bundle1, bundle2]);
     assert.deepEqual(
-      mapContextSync(contexts, ['foo', 'bar']),
-      [ctx1, ctx1]
+      mapBundleSync(bundles, ['foo', 'bar']),
+      [bundle1, bundle1]
     );
   });
 
   test('fallback translations', function() {
-    const contexts = new CachedSyncIterable([ctx1, ctx2]);
+    const bundles = new CachedSyncIterable([bundle1, bundle2]);
     assert.deepEqual(
-      mapContextSync(contexts, ['foo', 'bar', 'baz']),
-      [ctx1, ctx1, ctx2]
+      mapBundleSync(bundles, ['foo', 'bar', 'baz']),
+      [bundle1, bundle1, bundle2]
     );
   });
 
   test('missing translations', function() {
-    const contexts = new CachedSyncIterable([ctx1, ctx2]);
+    const bundles = new CachedSyncIterable([bundle1, bundle2]);
     assert.deepEqual(
-      mapContextSync(contexts, ['foo', 'bar', 'baz', 'qux']),
-      [ctx1, ctx1, ctx2, null]
+      mapBundleSync(bundles, ['foo', 'bar', 'baz', 'qux']),
+      [bundle1, bundle1, bundle2, null]
     );
   });
 });