+++ fluent/test/patterns_test.js
@@ -2,11 +2,11 @@
 
 import assert from 'assert';
 
-import { MessageContext } from '../src/context';
+import { FluentBundle } from '../src/context';
 import { ftl } from '../src/util';
 
 suite('Patterns', function(){
-  let ctx, args, errs;
+  let bundle, args, errs;
 
   setup(function() {
     errs = [];
@@ -14,15 +14,15 @@ suite('Patterns', function(){
 
   suite('Simple string value', function(){
     suiteSetup(function() {
-      ctx = new MessageContext('en-US', { useIsolating: false });
-      ctx.addMessages(ftl`
+      bundle = new FluentBundle('en-US', { useIsolating: false });
+      bundle.addMessages(ftl`
         foo = Foo
       `);
     });
 
     test('returns the value', function(){
-      const msg = ctx.getMessage('foo');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('foo');
+      const val = bundle.format(msg, args, errs);
       assert.equal(val, 'Foo');
       assert.equal(errs.length, 0);
     });
@@ -30,8 +30,8 @@ suite('Patterns', function(){
 
   suite('Complex string value', function(){
     suiteSetup(function() {
-      ctx = new MessageContext('en-US', { useIsolating: false });
-      ctx.addMessages(ftl`
+      bundle = new FluentBundle('en-US', { useIsolating: false });
+      bundle.addMessages(ftl`
         foo = Foo
         -bar = Bar
 
@@ -46,29 +46,29 @@ suite('Patterns', function(){
     });
 
     test('resolves the reference to a message', function(){
-      const msg = ctx.getMessage('ref-message');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('ref-message');
+      const val = bundle.format(msg, args, errs);
       assert.strictEqual(val, 'Foo');
       assert.equal(errs.length, 0);
     });
 
     test('resolves the reference to a term', function(){
-      const msg = ctx.getMessage('ref-term');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('ref-term');
+      const val = bundle.format(msg, args, errs);
       assert.strictEqual(val, 'Bar');
       assert.equal(errs.length, 0);
     });
 
     test('returns the id if a message reference is missing', function(){
-      const msg = ctx.getMessage('ref-missing-message');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('ref-missing-message');
+      const val = bundle.format(msg, args, errs);
       assert.strictEqual(val, 'missing');
       assert.ok(errs[0] instanceof ReferenceError); // unknown message
     });
 
     test('returns the id if a term reference is missing', function(){
-      const msg = ctx.getMessage('ref-missing-term');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('ref-missing-term');
+      const val = bundle.format(msg, args, errs);
       assert.strictEqual(val, '-missing');
       assert.ok(errs[0] instanceof ReferenceError); // unknown message
     });
@@ -76,8 +76,8 @@ suite('Patterns', function(){
 
   suite('Complex string referencing a message with null value', function(){
     suiteSetup(function() {
-      ctx = new MessageContext('en-US', { useIsolating: false });
-      ctx.addMessages(ftl`
+      bundle = new FluentBundle('en-US', { useIsolating: false });
+      bundle.addMessages(ftl`
         foo =
             .attr = Foo Attr
         bar = { foo } Bar
@@ -85,23 +85,23 @@ suite('Patterns', function(){
     });
 
     test('returns the null value', function(){
-      const msg = ctx.getMessage('foo');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('foo');
+      const val = bundle.format(msg, args, errs);
       assert.strictEqual(val, null);
       assert.equal(errs.length, 0);
     });
 
     test('formats the attribute', function(){
-      const msg = ctx.getMessage('foo');
-      const val = ctx.format(msg.attrs.attr, args, errs);
+      const msg = bundle.getMessage('foo');
+      const val = bundle.format(msg.attrs.attr, args, errs);
       assert.strictEqual(val, 'Foo Attr');
       assert.equal(errs.length, 0);
     });
 
     test('formats ??? when the referenced message has no value and no default',
        function(){
-      const msg = ctx.getMessage('bar');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('bar');
+      const val = bundle.format(msg, args, errs);
       assert.strictEqual(val, '??? Bar');
       assert.ok(errs[0] instanceof RangeError); // no default
     });
@@ -109,16 +109,16 @@ suite('Patterns', function(){
 
   suite('Cyclic reference', function(){
     suiteSetup(function() {
-      ctx = new MessageContext('en-US', { useIsolating: false });
-      ctx.addMessages(ftl`
+      bundle = new FluentBundle('en-US', { useIsolating: false });
+      bundle.addMessages(ftl`
         foo = { bar }
         bar = { foo }
       `);
     });
 
     test('returns ???', function(){
-      const msg = ctx.getMessage('foo');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('foo');
+      const val = bundle.format(msg, args, errs);
       assert.strictEqual(val, '???');
       assert.ok(errs[0] instanceof RangeError); // cyclic reference
     });
@@ -126,15 +126,15 @@ suite('Patterns', function(){
 
   suite('Cyclic self-reference', function(){
     suiteSetup(function() {
-      ctx = new MessageContext('en-US', { useIsolating: false });
-      ctx.addMessages(ftl`
+      bundle = new FluentBundle('en-US', { useIsolating: false });
+      bundle.addMessages(ftl`
         foo = { foo }
       `);
     });
 
     test('returns the raw string', function(){
-      const msg = ctx.getMessage('foo');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('foo');
+      const val = bundle.format(msg, args, errs);
       assert.strictEqual(val, '???');
       assert.ok(errs[0] instanceof RangeError); // cyclic reference
     });
@@ -142,8 +142,8 @@ suite('Patterns', function(){
 
   suite('Cyclic self-reference in a member', function(){
     suiteSetup(function() {
-      ctx = new MessageContext('en-US', { useIsolating: false });
-      ctx.addMessages(ftl`
+      bundle = new FluentBundle('en-US', { useIsolating: false });
+      bundle.addMessages(ftl`
         foo =
             { $sel ->
                *[a] { foo }
@@ -154,15 +154,15 @@ suite('Patterns', function(){
     });
 
     test('returns ???', function(){
-      const msg = ctx.getMessage('foo');
-      const val = ctx.format(msg, {sel: 'a'}, errs);
+      const msg = bundle.getMessage('foo');
+      const val = bundle.format(msg, {sel: 'a'}, errs);
       assert.strictEqual(val, '???');
       assert.ok(errs[0] instanceof RangeError); // cyclic reference
     });
 
     test('returns the other member if requested', function(){
-      const msg = ctx.getMessage('foo');
-      const val = ctx.format(msg, {sel: 'b'}, errs);
+      const msg = bundle.getMessage('foo');
+      const val = bundle.format(msg, {sel: 'b'}, errs);
       assert.strictEqual(val, 'Bar');
       assert.equal(errs.length, 0);
     });
@@ -170,8 +170,8 @@ suite('Patterns', function(){
 
   suite('Cyclic reference in a selector', function(){
     suiteSetup(function() {
-      ctx = new MessageContext('en-US', { useIsolating: false });
-      ctx.addMessages(ftl`
+      bundle = new FluentBundle('en-US', { useIsolating: false });
+      bundle.addMessages(ftl`
         -foo =
             { -bar.attr ->
                *[a] Foo
@@ -184,8 +184,8 @@ suite('Patterns', function(){
     });
 
     test('returns the default variant', function(){
-      const msg = ctx.getMessage('foo');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('foo');
+      const val = bundle.format(msg, args, errs);
       assert.strictEqual(val, 'Foo');
       assert.ok(errs[0] instanceof RangeError); // cyclic reference
     });
@@ -193,8 +193,8 @@ suite('Patterns', function(){
 
   suite('Cyclic self-reference in a selector', function(){
     suiteSetup(function() {
-      ctx = new MessageContext('en-US', { useIsolating: false });
-      ctx.addMessages(ftl`
+      bundle = new FluentBundle('en-US', { useIsolating: false });
+      bundle.addMessages(ftl`
         -foo =
             { -bar.attr ->
                *[a] Foo
@@ -213,15 +213,15 @@ suite('Patterns', function(){
     });
 
     test('returns the default variant', function(){
-      const msg = ctx.getMessage('foo');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('foo');
+      const val = bundle.format(msg, args, errs);
       assert.strictEqual(val, 'Foo');
       assert.ok(errs[0] instanceof RangeError); // cyclic reference
     });
 
     test('can reference an attribute', function(){
-      const msg = ctx.getMessage('bar');
-      const val = ctx.format(msg, args, errs);
+      const msg = bundle.getMessage('bar');
+      const val = bundle.format(msg, args, errs);
       assert.strictEqual(val, 'Bar');
       assert.equal(errs.length, 0);
     });