+++ tools/perf/benchmark.d8.js
@@ -19,15 +19,15 @@ times.ftlEntriesParseStart = Date.now();
 var [entries] = Fluent._parse(ftlCode);
 times.ftlEntriesParseEnd = Date.now();
 
-var ctx = new Fluent.MessageContext('en-US');
-var errors = ctx.addMessages(ftlCode);
+var bundle = new Fluent.FluentBundle('en-US');
+var errors = bundle.addMessages(ftlCode);
 
 times.format = Date.now();
-for (const [id, message] of ctx.messages) {
-  ctx.format(message, args, errors);
+for (const [id, message] of bundle.messages) {
+  bundle.format(message, args, errors);
   if (message.attrs) {
     for (const name in message.attrs) {
-      ctx.format(message.attrs[name], args, errors)
+      bundle.format(message.attrs[name], args, errors)
     }
   }
 }