+++ tools/perf/benchmark.node.js
@@ -23,15 +23,15 @@ cumulative.ftlEntriesParseStart = process.hrtime(start);
 var [entries] = Fluent._parse(ftlCode);
 cumulative.ftlEntriesParseEnd = process.hrtime(start);
 
-var ctx = new Fluent.MessageContext('en-US');
-var errors = ctx.addMessages(ftlCode);
+var bundle = new Fluent.FluentBundle('en-US');
+var errors = bundle.addMessages(ftlCode);
 
 cumulative.format = process.hrtime(start);
-for (const [id, message] of ctx.messages) {
-  ctx.format(message, args, errors);
+for (const [id, message] of bundle.messages) {
+  bundle.format(message, args, errors);
   if (message.attrs) {
     for (const name in message.attrs) {
-      ctx.format(message.attrs[name], args, errors)
+      bundle.format(message.attrs[name], args, errors)
     }
   }
 }